SQLGeneration/pyproject.toml
[project]
name = "sqlgeneration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=40.15.0",
    "fastapi>=0.136.0",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.16",
    "langgraph>=1.1.9",
    "mcp[cli]>=1.27.0",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "pymongo>=4.17.0",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "tiktoken>=0.12.0",
    "uvicorn>=0.45.0",
]
